Use Date.timeIntervalSince1970 to create Coordinator identifier
Right now, to create a flow coordinator's identifier we use (FlowName)-(Date). Which prints something similar to "AppFlow-2018-08-19 19:49:40 +0000".
This only gives us granularity up to the second and does not create unique identifiers for multiple Flows of the same type being created within the same second.
Because the data structure that holds all the flowCoordinators' references is a Dictionary, when multiple flows of the same type are created at the same time, when we try to set them in the dictionary, only one flowCoordinator will be stored and the other references will be lost. This causes the Step events to be disposed.

This commit uses timeIntervalSince1970, which gives us sub-millisecond granularity.