flow: add a filtering feature before navigate(to:)
A filter(step:) -> Single<Step> function now allows to filter and adapt
an incoming step within a Flow before it is given to the navigate(to:) function.

One possible usecase is the permission management in an application: the filter
function could forbid a navigation if all the required permissions are not granted.
The Demo App has been amended to implement such a usecase.