lint: enable file_name rule to fix warning
Before this commit, there was a conflict between swiftlint rules, which
produced a "Found a configuration for 'file_name' rule, but it is
disabled on 'disabled_rules'." warning.

This commit removes the conflict by removing the line in disable, making
the one in opt_in active, which will then use the given .swiftlint.yml
configuration.

Doing so, some files now raise a warning because of their filename.